pyproject.toml
[project]
name = "wbia-plugin-miew-id"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "albumentations==1.3.0",
    "grad-cam==1.4.6",
    "ipykernel>=6.29.5",
    "jupyterlab>=4.3.8",
    "numpy",
    "opencv-python-headless==4.7.0.72",
    "optuna==3.2.0",
    "pandas==1.3.5",
    "pillow==9.5.0",
    "pip",
    "python-dotenv==1.0.0",
    "pyyaml==6.0",
    "scikit-learn==1.2.2",
    "scipy==1.7.2",
    "timm==0.9.2",
    "torch",
    "torchvision",
    "tqdm==4.65.0",
    "transformers>=4.56.1",
    "uv>=0.8.17",
    "wandb>=0.21.4",
    "zennit==0.5.1",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu124", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu124", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
